time dillation p1

+++ main.cpp
@@ -29,8 +29,8 @@ int main() {
 
 	Game game;
 	Texture2D texture = LoadTexture("./resources/textures/adventurer.png");
-	//            pos X    pos Y     width of texture  height of texture  width             height          texture
-	Player player(WIDTH/2, HEIGHT/3, texture.width/21, texture.height,    texture.width/21, texture.height, texture);
+	//            pos X    pos Y         width of texture  height of texture  width             height          texture
+	Player player(float(WIDTH/2), float(HEIGHT/3),    float( texture.width/21), float(texture.height),    float(texture.width/21), float(texture.height), texture);
 	Rectangle floor  = { -100, HEIGHT/2 + 40, 10000, 50 };
 
 	std::vector<Rectangle> objects;
@@ -43,7 +43,6 @@ int main() {
 
 	Camera2D camera = {{ 0 }, {0, 0}, 0.0f, 0.0f };
 	camera.zoom = 1.0f;
-
 	SetTargetFPS(60);
 
 	while (!WindowShouldClose()){
